DaanCoronaBackend/requirements.txt
Django==2.2.11
django-otp==0.8.1
django-rest-knox==2.3.0
django-rest-swagger==2.1.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djangorestframework-yaml==1.0.3
djoser==2.0.3
Pillow==6.2.1
PyJWT==1.7.1
python-decouple==3.3
requests==2.13.0
requests-oauthlib==0.3.1
twilio==6.37.0
boto3==1.12.31
